Bug fix in C interface

+++ src/proto.h
@@ -37,7 +37,7 @@ void Clinvwhichdist(int *, int *, double *, int *, int *, int *, int *, double *
 void linvknndist(int *, int *, int *, double *, int *, int *, int *, int *, double *, double *, double *, double *, int *);
 void linSpairUdist(int *, int *, double *, int *, int *, int *, int *, double *, double *, double *, double *); 
 void linSpairdist(int *, int *, double *, int *, int *, int *, int *, double *, double *, double *, double *); 
-void linpairdist(int *, double *, double *, int *, double *, double *, double *, int *, int *, double *, int *, double *);
+void linpairdist(int *, double *, double *, int *, double *, double *, int *, int *, int *, double *, int *, double *);
 /*
 
              Functions invoked by .Call
